+++ environment
@@ -126,7 +126,7 @@ DSS_INFRA_TAG_OWNER="chmreid@ucsc.edu"
 DSS_NOTIFICATION_SENDER="chmreid@ucsc.edu"
 API_DOMAIN_NAME="dss.${DSS_DEPLOYMENT_STAGE}.ucsc-cgp-redwood.org"
 
-TOKENINFO_FUNC=dss.util.auth.helpers.verify_jwt
+TOKENINFO_FUNC=dss.util.security.verify_jwt
 # TODO remove https://dev.data.humancellatlas.org/ from OIDC_AUDIENCE once seperate deployments are made
 
 OIDC_AUDIENCE=https://dss.dev.ucsc-cgp-redwood.org
